Bato.to: Refactor mirror and version selection
* Merge mirror lists from V2 and V4 into a single preference.
* Remove manual site version selector.
* Automatically switch between V2 and V4 implementation based on the selected mirror.
* Centralize base URL handling.
* Switch delegated source without need to restart
